pyproject.toml
[project]
name = "nutrition-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.7.1.1",
    "faiss-cpu>=1.12.0",
    "google-generativeai>=0.8.5",
    "gradio>=5.43.1",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-groq>=0.3.7",
    "langchain-postgres>=0.0.15",
    "langgraph>=0.6.6",
    "loguru>=0.7.3",
    "openai>=1.101.0",
    "protobuf>=5.29.5",
    "psycopg[binary]>=3.2.9",
    "psycopg-binary>=3.2.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
    "aiofiles>=24.1.0",
    "ragas>=0.3.2",
    "langsmith",
]
